Move RecordingHiveMetastore to recording package
